requirements.txt
discord.py==2.3.2
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
apscheduler==3.10.4
python-dateutil==2.9.0.post0
pytz==2024.1
python-dotenv==1.0.1
